feat(laufen): swap clerc → @kidd-cli/core for the CLI surface
Rebuilds the laufen CLI on top of @kidd-cli/core (kidd, soon-to-be maltty)
in place of clerc + the home-grown handler shim. Result is dramatically
less plumbing — every command now uses kidd's command()/ctx form with
ctx.fail / ctx.log / ctx.prompts / ctx.colors directly.

New structure:
- src/commands/{init,list,info,run,create,blueprint}.ts — one file per
  command using kidd's command() with Zod options/positionals
- src/index.ts — kidd cli() bootstrap that wires the commands map
- src/lib/result.ts — keeps the Result tuple internally + adds an
  assertOk() helper that lifts kidd's ctx.fail into a typed assertion,
  so `assertOk(result, ctx.fail, '...')` narrows the value side

Deleted (replaced by kidd):
- src/handlers/** (6 handler files + tests)
- src/lib/handler.ts + test — defineHandler shim
- src/utils/prompt.ts + test — promptForScript/promptForText
- src/utils/prompt-args.ts + test — the local copy of script-arg
  prompting (the engine has its own copy that the executor still uses)
- src/utils/schema.ts — JSON-schema introspection used only by prompt-args
- src/utils/help-rewrite.ts — clerc-specific help argv rewriting
- src/utils/resolve-script.ts — folded into commands/{info,run}.ts
- src/state/script-help.ts — clerc-side help flag mailbox

src/utils/cli.ts: trimmed to safeParseError + readPackageJSON
(formatArgErrors and errorHint were clerc-only).

Lint config: added commands/** to the override that turns off the
functional/no-* rules for handler-style code.

Validation: lint / fmt:check / typecheck all green. Tests for the new
commands and an end-to-end smoke run are intentionally deferred to the
next commit on this branch.
