+++ packages/laufen/src/index.ts
@@ -1,106 +1,37 @@
 #!/usr/bin/env node
-// oxlint-disable import/max-dependencies
-
-import path from 'node:path';
 import { fileURLToPath } from 'node:url';
 
-import * as p from '@clack/prompts';
-import { Clerc, completionsPlugin, helpPlugin, versionPlugin } from 'clerc';
-import pc from 'picocolors';
+import { cli } from '@kidd-cli/core';
+import * as path from 'pathe';
 
-import handleBlueprint from './handlers/blueprint.ts';
-import handleCreate from './handlers/create.ts';
-import handleInfo from './handlers/info.ts';
-import handleInit from './handlers/init.ts';
-import handleList from './handlers/list.ts';
-import handleRun from './handlers/run.ts';
-import { markScriptHelpRequested } from './state/script-help.ts';
-import { errorHint, readPackageJSON, safeParseError } from './utils/cli.ts';
-import { rewriteHelpArgv } from './utils/help-rewrite.ts';
+import blueprint from './commands/blueprint.ts';
+import create from './commands/create.ts';
+import info from './commands/info.ts';
+import init from './commands/init.ts';
+import list from './commands/list.ts';
+import run from './commands/run.ts';
+import { readPackageJSON } from './utils/cli.ts';
 
 const pkgDir = path.dirname(path.dirname(fileURLToPath(import.meta.url)));
 const [pkgError, pkg] = readPackageJSON(pkgDir);
 
 if (pkgError || pkg === null || !pkg.version) {
-  p.log.error(
-    'Fatal error, unable to execute lauf, please log an issue on github: https://github.com/zrosenbauer/lauf/issues',
+  process.stderr.write(
+    'Fatal error, unable to execute lauf, please log an issue on github: https://github.com/zrosenbauer/lauf/issues\n',
   );
   process.exit(1);
 }
 
-const { argv: rewrittenArgv, scriptHelpRequested } = rewriteHelpArgv(process.argv.slice(2));
-
-if (scriptHelpRequested) {
-  markScriptHelpRequested();
-}
-
-Clerc.create()
-  .name('lauf')
-  .scriptName('lauf')
-  .description('Typed script runner for monorepos')
-  .version(pkg.version)
-  .use(helpPlugin())
-  .use(versionPlugin())
-  .use(completionsPlugin())
-  .errorHandler((err: unknown) => {
-    const message = safeParseError(err);
-    p.log.error(message);
-    const hint = errorHint(err);
-    if (hint) {
-      p.log.message(pc.dim(hint));
-    }
-    process.exit(1);
-  })
-  .command('init', 'Initialize lauf in the current package')
-  .on('init', handleInit)
-  .command('list', 'List all available scripts', {
-    flags: {
-      all: {
-        type: Boolean,
-        description: 'Discover scripts from all nested configs',
-        alias: 'a',
-      },
-      filter: {
-        type: String,
-        description: 'Filter packages by name glob (e.g. "@apps/*")',
-        alias: 'f',
-      },
-    },
-  })
-  .on('list', handleList)
-  .command('info', 'Show info for a script', {
-    parameters: ['[script]'],
-  })
-  .on('info', handleInfo)
-  .command('run', 'Run a script', {
-    parameters: ['[script]'],
-    flags: {
-      watch: {
-        type: Boolean,
-        description: 'Watch files and rerun on changes',
-        alias: 'w',
-      },
-    },
-  })
-  .on('run', handleRun)
-  .command('create', 'Create a new script', {
-    parameters: ['[name]'],
-    flags: {
-      dir: {
-        type: String,
-        description: 'Target directory (relative to monorepo root)',
-      },
-    },
-  })
-  .on('create', handleCreate)
-  .command('blueprint', 'Scaffold a pre-built script blueprint', {
-    parameters: ['[name]'],
-    flags: {
-      dir: {
-        type: String,
-        description: 'Target directory (relative to monorepo root)',
-      },
-    },
-  })
-  .on('blueprint', handleBlueprint)
-  .parse([...rewrittenArgv]);
+await cli({
+  name: 'lauf',
+  version: pkg.version,
+  description: 'Typed script runner for monorepos',
+  commands: {
+    init,
+    list,
+    info,
+    run,
+    create,
+    blueprint,
+  },
+});
