+++ packages/laufen/src/utils/cli.test.ts
@@ -1,9 +1,8 @@
 import * as fs from 'node:fs';
 
-import { InvalidParametersError, MissingRequiredFlagError, NoSuchCommandError } from 'clerc';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
-import { errorHint, formatArgErrors, readPackageJSON, safeParseError } from './cli.ts';
+import { readPackageJSON, safeParseError } from './cli.ts';
 
 vi.mock('node:fs', () => ({
   readFileSync: vi.fn(),
@@ -45,68 +44,6 @@ describe('safeParseError', () => {
 
   it('extracts message from Error subclass', () => {
     expect(safeParseError(new TypeError('type error'))).toBe('type error');
-  });
-});
-
-describe('formatArgErrors', () => {
-  it('formats a single issue', () => {
-    const issues = [{ path: ['name'], message: 'Required', code: 'invalid_type' }];
-    const result = formatArgErrors(issues as never);
-    expect(result).toContain('Invalid arguments:');
-    expect(result).toContain('--name: Required');
-  });
-
-  it('formats multiple issues', () => {
-    const issues = [
-      { path: ['name'], message: 'Required', code: 'invalid_type' },
-      { path: ['count'], message: 'Expected number', code: 'invalid_type' },
-    ];
-    const result = formatArgErrors(issues as never);
-    expect(result).toContain('--name: Required');
-    expect(result).toContain('--count: Expected number');
-  });
-
-  it('handles nested paths', () => {
-    const issues = [{ path: ['config', 'output'], message: 'Invalid', code: 'invalid_type' }];
-    const result = formatArgErrors(issues as never);
-    expect(result).toContain('--config.output: Invalid');
-  });
-
-  it('handles empty issues array', () => {
-    const result = formatArgErrors([]);
-    expect(result).toBe('Invalid arguments:\n');
-  });
-});
-
-describe('errorHint', () => {
-  it('returns help hint for InvalidParametersError', () => {
-    const err = new InvalidParametersError('test');
-    const hint = errorHint(err);
-    expect(hint).toBe('Run `lauf <command> --help` for usage information.');
-  });
-
-  it('returns help hint for MissingRequiredFlagError', () => {
-    const err = new MissingRequiredFlagError(['test']);
-    const hint = errorHint(err);
-    expect(hint).toBe('Run `lauf <command> --help` for usage information.');
-  });
-
-  it('returns command hint for NoSuchCommandError', () => {
-    const err = new NoSuchCommandError('test');
-    const hint = errorHint(err);
-    expect(hint).toBe('Run `lauf --help` to see available commands.');
-  });
-
-  it('returns undefined for generic Error', () => {
-    expect(errorHint(new Error('generic'))).toBeUndefined();
-  });
-
-  it('returns undefined for string error', () => {
-    expect(errorHint('string error')).toBeUndefined();
-  });
-
-  it('returns undefined for null', () => {
-    expect(errorHint(null)).toBeUndefined();
   });
 });
