--- packages/laufen/src/utils/schema.ts
@@ -1,2 +1,0 @@
-export type { JsonSchemaProperty } from '@laufen/engine';
-export { extractSchemaFields, resolveType } from '@laufen/engine';
